% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cppRIT.R
\name{relaxed_RIT}
\alias{relaxed_RIT}
\title{Relaxed Random Intersection Trees.}
\usage{
relaxed_RIT(data, classes, theta, epsilon, epsilon_cat, n_trees = 100L,
  depth = 10L, min_inter_sz = 2L, branch = 5, es = TRUE)
}
\arguments{
\item{data}{The dataset as a dataframe.}

\item{classes}{A response vector for the dataset.}

\item{theta}{The prevalence thresholds. Used to determine if an interaction is informative of not w.r.t. the classes.}

\item{epsilon}{The scaling factor applied to the cumulative error threshold for continuous features.}

\item{epsilon_cat}{The scaling factor applied to the cumulative error threshold for categorical features.}

\item{n_trees}{The number of trees.}

\item{depth}{The depth of the interaction search.}

\item{min_inter_sz}{The minimum allowed size of interactions.}

\item{branch}{The branching factor in the interaction search.}

\item{es}{A logical that indicates whether or not early stopping should be used.}
}
\value{
A model that is composed of all informative interactions found in the dataset.
}
\description{
Executes the algorithm known as Relaxed Random Intersection Trees.
}
\references{
Ballarini Robin. Random intersection trees for genomic data analysis. Master's thesis, Université Catholique de Louvain, 2016.
}

