% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RIT_predict.R
\name{rit_predict}
\alias{rit_predict}
\title{Classification Rule for Relaxed Random Intersection Trees.}
\usage{
rit_predict(rit, testset)
}
\arguments{
\item{rit}{A model produced by \code{relaxed_RIT}}

\item{testset}{A dataframe containing the instances to classify}
}
\value{
A response vector for the \code{testset} instances
}
\description{
Applies a basic \code{argmax} rule in order to classify new instances.
}
\references{
Ballarini Robin. Random intersection trees for genomic data analysis. Ecole polytechnique de Louvain, Université catholique de Louvain, 2016. Prom. : Dupont, Pierre.
}

